Переносит моковые данные в отдельный файл
