+++ src/mock.js
@@ -1,0 +1,12 @@
+const promoInfo = {
+  title: `The Grand Budapest Hotel`,
+  genre: `Drama`,
+  releaseDate: 2014
+};
+
+const moviesNames = [`Fantastic Beasts: The Crimes of Grindelwald`, `Bohemian Rhapsody`, `Macbeth`, `Aviator`, `We need to talk about Kevin`, `We need to talk about Kevin`, `Revenant`, `Johnny English`, `Shutter Island`, `Pulp Fiction`, `No Country for Old Men`, `Snatch`, `Moonrise Kingdom`, `Seven Years in Tibet`, `Midnight Special`, `War of the Worlds`, `Dardjeeling Limited`, `Orlando`, `Mindhunter`, `Midnight Special`];
+
+export {
+  promoInfo,
+  moviesNames,
+};
